Disable tiered code cache on z Systems
Tiered code cache has always been a dubious optimization to avoid paying
instruction cache footprint when branching between methods, however no
measurable benefit can be seen from this optimization. With the
introduction of long relative branches this optimization is moot.

With tiered code caches enabled we may actually end up allocating more
code cache than the default. The reason for this is because of the way
we generate warm binary code (forward from the top of a code cache) and
cold binary code (backward from the bottom of the code cache). With the
backward allocation of cold code caches we are not able to reclaim the
over-estimation in the size of the code cache required due to the
variable length instructions used on z Systems. Because of this we end
up wasting precious code cache space and exhibiting a performance
degradation.

Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>