Update append instruction when deleting the old
The append instruction, which is the instruction to which new
instructions will be appended to in the instruction linked list, should
be updated when the old append instruction is being removed.

This avoids a scenario in which the append instruction pointer is not
updated, and all subsequent generated instruction are not binary encoded
because they are not referenced from the linked list of the instruction
sequence starting from the beginning of the compiled body.

Issue: #212
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>