Prevent invalid memory copy in directStoreHelper
It is invalid to perform a memory copy transformation if there is a
non-integral truncation conversion in between the load and the store
nodes. The reason being that the types could be wildly different and
that the conversion is non-trivial. Such an example occurs the the
following tree:

istore
  d2i
    dload

For which the d2i conversion is non-trivial albeit technically a
truncation. We must prohibit this memory copy optimization for non
integral conversions.

Task 118388

Issue: #205
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>