Trace when an opt pass stops due to getNodeCount()
A few optimization passes stop in their tracks, or avoid starting in the
first place, when `getNodeCount()` (which is really the next available
global index) is above some threshold. Until now dead trees elimination
and value propagation could do this silently, which would look alarming
in a trace log.

These optimizations now log a message when abandoning the pass in this
way, and dead trees elimination also prints a similar message if it
gives up due its visit count limit.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>