Replace _tryCatchInfo implementation with a deque
Change the implementation of the _tryCatchInfo container to use a deque
instead of a CS2::ArrayOf.  To that end, provide a constructor for the
struct's definition, and use insert(begin() + index) rather than relying
on the [] indexing operator to grow the container.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>
(cherry picked from commit aa39c25d9847ddcb430fd7b7548f57d66e21567f)