Test vlog creation and limitfile adherence
Limitfiles are a way to whitelist which methods are compiled by the
JIT. This can be useful in determining which function is triggering
a compiler bug.

Verbose log (vlog) files are logs that can be generated by the
JIT. Among other things, these include information about which
functions have been compiled. These files can be used as limitfiles.

These new tests test that vlog files are successfully created, and
that they can be used as limitfiles by the JIT.

This also runs some basic tests to ensure the JIT uses the limitfile
as a whitelist, using either a part of the limitfile or the entire
limitfile.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>
(cherry picked from commit f897ed1e2a4aa51aba528b43e7a017c4fce22529)