Replace remaining outOfMemory with exceptions
This commit removes all calls to outOfMemory that was not already
removed in a previous commit and replaces it with a traceMsg and a C++
exception. This commit completes Step 1 of issue 380 (at least in the
OMR project).

Issue: #380
Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>
(cherry picked from commit 8287e54286580f4d921c1a295eac42e1cd2c9412)