Remove replaceWithDestructiveFormIfPossible
Currently this function exists to transform instructions of the form:

    ARK Rx, Ry, Rz # (Rx = Ry + Rz)

to

    AR  Rx, Rz     # (Rx += Rz)

when x == y.

However, there are no cases of it being used. To
complicate matters, this code requires a member variable implemented as
a CS2::ArrayOf, which is currently being targetted for refactoring and
elimination.

Rather than refactor this dead code, this change removes it from current
source.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>
(cherry picked from commit 25f1cbcbec6305a25ecc7621928de7662341c3b0)