Stop IVA rejecting tests of loads with refcount 1
IVA (induction variable analysis) has been rejecting primary induction
variable candidates whenever the loop test is expressed as a comparison
against a load of the variable, and the load has reference count 1. Such
loop tests would be subjected to the requirements for non-load
expressions: that there be a store to the variable earlier in the same
extended block, and that the expression tested be the right-hand side of
the last such store. But the load can't possibly appear directly under a
store, because if it did, it would have reference count >=2.

When analyzing the loop test, IVA now determines whether a load could
possibly be evaluated earlier than the loop test tree, based on the
reference count of the load, and the reference count of its parent
conversion operation if any. If a load has definitely *not* been
evaluated before the test, then that load produces the updated value,
and there is no need to reject the test. The handling for loads with
reference count >1 is extended to apply to any load that might be
evaluated before the test, i.e. to loads with reference count 1 that
were found beneath a conversion with reference count >1.

Signed-off-by: Devin Papineau <devinmp@ca.ibm.com>
(cherry picked from commit 8ea5bba1179481eb7d28933619e1757a460ed75b)