Add template method to fail compilation
Before the outOfMemory calls were replaced with exception throws, in
addition to writing a message to the trace log it would also print the
message to stderr. This commit adds a template method in the
OMR::Compilation class called failCompilation which will print out the
message to the trace log as well as stderr under an option, and will
then throw an exception as specified by the template parameter.

Issue: #380
Signed-off-by: Irwin D'Souza <dsouzai@ca.ibm.com>
(cherry picked from commit 251781595442523d1f3e26b62e30a806cd9d7fc3)