Fix bug with bytecode builders with no fallthrough
Fix JitBuilder bug where when a method only has BytecodeBuilder objects
with not fallthrough path, non-connected/unreachable builders are not
added to the end of the tree list as they should.

The problem is resolved by changing the outermost while loop in
MethodBuilder::connectTrees to a do-while loop. This ensures that the
inner loop responsible for adding unreachable builders to the work list
is executed at least once. This is needed because, when the only
reachable builder has no fallthrough builder, no builder objects are
added to the work list, which means the list is initially empty, which
means the outer while loop never gets executed. Hence, the unreachable
builders never get appended to the list.

Issue: #488

Signed-off-by: Leonardo Banderali <leob@linux.vnet.ibm.com>
(cherry picked from commit b43cfa5b01b84067611be04211392f360828ca34)