Remove `ifInstanceOfHelper`
The method `OMR::X86::TreeEvaluator::ifInstanceOfHelper` simply
called another function with the same parameters, and was only called
in two locations. If this is required in the future, it can simply be
recreated.

The method `OMR::X86::AMD64::TreeEvaluator::ifInstanceOfHelper` was
not implemented. The calls to the `OMR::X86` version did not use
polymorphism, so this was never invoked.

Signed-off-by: Lukas Lorimer <lukasl@ca.ibm.com>
(cherry picked from commit 471be23798682cae2b03444da36863b46b1039bb)