Use a deque for collecting internal labels
The isInternalControlFlowOneEntryOneExit helper function in the z
version of OMRInstruction.cpp uses a CS2::ArrayOf local to accumulate
known labels for use in determining whether or not branch instructions'
targets are either local or non-local.  This change set replaces that
container with a TR::deque, and changes the iteration of the container
from using an integer index to using iterators.  Finally, this change
adds a break statement if we find a matching label to avoid needlessly
continuing through the container after we have a confirmed match.

Signed-off-by: Luc des Trois Maisons <lmaisons@ca.ibm.com>
(cherry picked from commit a5d1ff61a55fa2e55fa2e0aa4c13906ce94c2bce)