Use failCompilation method to fail compilations
This commit uses the failCompilation template method to fail
compilations instead of explicitly writing to the trace log and throwing
an exception. The type of the exception is passed as a template
parameter and the method handles printing the message to both the
tracelog and stderr if necessary.

Issue: #380
Signed-off-by: Irwin D'Souza <dsouza@ca.ibm.com>
(cherry picked from commit 7c9d37df93cfee6eab0bda9c634cfb9b1e88d1f3)