Fix LoopVersioner to treat receivers outside of the loop as variant
When LoopVersioner decides whether a guard is loop invariant, it looks
at the receiver on the taken side as the last resort. In some cases,
the taken side is out of the loop and hard to reason about. The
LoopVersioner is fixed to conservatively treat those receivers as
variant.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>
(cherry picked from commit b792785a41fa85c3a71c7c3acdb9bf677144505e)