Load Power Table Of Constants in method prologue
The Power Table of Constants (TOC) was not being loaded into r2 as
specified by the Power System Linkage. When the caller does not do this,
the callee will try to use the TOC base register, r2, resulting in an
error. To remedy this, the TOC will be loaded when the prologue is
created.

Signed-off-by: Ryan Santhirarajan <rsanth@ca.ibm.com>
(cherry picked from commit a3603e7194029755206dd39f257d96704f8f7354)