Resolve issue where the operand stack commits to the wrong address
When the VirtualMachineOperandStack *copy* constructor is invoked it
does not copy the _stackOffset variable.  This causes commits to the
actual stack to happen at the incorrect location. The solution is to
copy the value of _stackOffset to the new VirtualMachineOperandStack

Signed-off-by: Charlie Gracie <charlie.gracie@gmail.com>
(cherry picked from commit 97eeb63815d5e6d1ddff0e44696f0b54713d349b)