Detect OSRCatchBlock exit edge in IndVarConversion
Need to bail out IndVarConversion if there is any exit edge
from inside the loop to an OSRCatchBlock. This is because
edge splitting is needed on every exit to shrink the induction
variable back but it is not safe to split edge to an OSRCatchBlock.

Signed-off-by: Yi Zhang <yizhang@ca.ibm.com>
(cherry picked from commit 1c8f2d9479dd1103517cd29398244d69c8cb3974)