Fix OperandStackTests test for JitBuilder
The type of the SP register used in one of the two OperandStackTests
tests was incorrect, and the test incorrectly tried to configure its
VirtualMachineOperandStack to use a stackOffset of 0. Correct the
stack to use the default stack offset of -1 (which just means the
stack pointer always points at the top element and so, initially,
the stack pointer is pointing one below the base of the memory
where the stack is held). That meant I had to update the verifyStack
function that was testing the difference between the top of the stack
and the base of the stack so that it properly understood the current
size of the stack. I also updated the output of the test to explicitly
mention that tests failed if they didn't all pass (it used to only
print out "ALL TESTS PASSED" if they all passed, but didn't print
anything if not all tests passed.

Issue: 629

Signed-off-by: Mark Stoodley <mstoodle@ca.ibm.com>
(cherry picked from commit 81e39872c8c8a35cef78a86d0f5b472459ffdd4d)