Fix ternaryEvaluator syntax error
An artifact was introduced at the end of the line that was not caught by
TravisCI builds. This is a syntax error which will result in compile
time failures on z Systems. This commit fixes the syntactical issues.

Issue: #667
Signed-off-by: Filip Jeremic <fjeremic@ca.ibm.com>
(cherry picked from commit 9b4775298433adb0bdbd62c6ec166f21cdf63380)